backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.12
python-multipart==0.0.6
boto3==1.28.90
aiofiles==23.1.0
openai>=1.0.0
python-dotenv==1.0.0
